refactor: typography types
