+++ projects/ion/src/lib/core/types/index.ts
@@ -25,3 +25,4 @@ export * from './tab-group';
 export * from './table';
 export * from './tag';
 export * from './tooltip';
+export * from './typography';
