+++ projects/ion/src/lib/core/types/typography.ts
@@ -1,0 +1,8 @@
+export type HeadingType = 'h1' | 'h2' | 'h3' | 'h4';
+export type HeadingWeight = 'medium' | 'bold';
+export type ColorScheme =
+  | 'primary'
+  | 'secondary'
+  | 'dark-primary'
+  | 'dark-secondary';
+export type HeadingSize = 'small' | 'medium' | 'normal';
