+++ projects/ion/src/lib/typography/heading/heading.component.ts
@@ -5,16 +5,13 @@ import {
   Input,
   ViewChild,
 } from '@angular/core';
+import {
+  ColorScheme,
+  HeadingSize,
+  HeadingType,
+  HeadingWeight,
+} from '../../core/types/typography';
 import { createElement } from '../../utils';
-
-export type HeadingType = 'h1' | 'h2' | 'h3' | 'h4';
-export type HeadingWeight = 'medium' | 'bold';
-export type ColorScheme =
-  | 'primary'
-  | 'secondary'
-  | 'dark-primary'
-  | 'dark-secondary';
-export type HeadingSize = 'small' | 'medium' | 'normal';
 
 @Component({
   selector: 'ion-heading',
